requirements.txt
Django==2.2
pytz==2018.9
django-sendgrid-v5==0.8.0
celery==4.1.1
https://github.com/formatcom/django-celery-beat/zipball/v1.1.1#egg=django-celery-beat
https://github.com/formatcom/django-celery-results/zipball/v1.1.0#egg=django-celery-results
django-celerybeat-status==0.0.10
static==1.1.1
static3==0.7.0
dj-static==0.0.6
django-admin-rangefilter==0.3.12
uWSGI==2.0.18
psycopg2==2.8.4
psycopg2-binary==2.8.4
django-cors-headers==2.1.0
djangorestframework==3.9.3
djangorestframework-gis==0.14
xlrd==1.2.0
xlwt==1.3.0
asgi-redis==1.4.2
redis==2.10.5
channels==1.1.5
daphne==1.3.0
postgis==1.0.4
django-rest-swagger==2.2.0
django-rest-auth==0.9.5
kombu==4.3.0
Pillow==6.2.1
django-extensions
django-rest-recaptcha==1.0.2
django-environ==0.4.5
